src/main/java/com/saifee/recipe/conveters/NotesCommandToNotes.java
package com.saifee.recipe.conveters;

import com.saifee.recipe.command.NotesCommand;
import com.saifee.recipe.domain.Notes;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class NotesCommandToNotes implements Converter<NotesCommand, Notes> {

    @Synchronized
    @Nullable
    @Override
    public Notes convert(NotesCommand source) {
        if(source == null) {
            return null;
        }
        final Notes notes = new Notes();
        notes.setId(source.getId());
        notes.setDescription(source.getDescription());
        return notes;
    }
}
